ofmcc-851 - refactor application view
