+++ frontend/src/views/applications/ApplicationsView.vue
@@ -5,7 +5,7 @@
       <div v-if="isEmpty(applications)">You have no applications on file</div>
       <v-data-table v-else :headers="headers" :items="applications" item-key="applicationId" density="compact">
         <template #item.actions="{ item }">
-          <router-link :to="getRoutingPath(item?.applicationId)">
+          <router-link :to="{ name: 'facility-details', params: { applicationGuid: item?.applicationId } }">
             {{ getApplicationAction(item) }}
           </router-link>
         </template>
@@ -25,20 +25,18 @@
         </template>
       </v-data-table>
     </v-skeleton-loader>
-    <CancelApplicationDialog :show="showCancelDialog" :applicationId="cancelledApplicationId" @close="toggleCancelDialog" />
+    <CancelApplicationDialog :show="showCancelDialog" :applicationId="cancelledApplicationId" @close="toggleCancelDialog" @cancel="cancelApplication" />
     <AppButton id="back-home-button" :primary="false" size="large" width="200px" :to="{ name: 'home' }">&larr; Back to Home</AppButton>
   </v-container>
 </template>
 
 <script>
-import { mapActions, mapState } from 'pinia'
 import AppButton from '@/components/ui/AppButton.vue'
-import { useApplicationsStore } from '@/stores/applications'
 import alertMixin from '@/mixins/alertMixin'
 import { isEmpty } from 'lodash'
 import format from '@/utils/format'
 import CancelApplicationDialog from '@/components/applications/CancelApplicationDialog.vue'
-import { PATHS, getApplicationUrl } from '@/services/routingService'
+import ApplicationService from '@/services/applicationService'
 
 export default {
   components: { AppButton, CancelApplicationDialog },
@@ -46,6 +44,7 @@ export default {
   data() {
     return {
       format,
+      applications: [],
       headers: [
         { title: 'Application ID', key: 'referenceNumber', width: '15%' },
         { title: 'Facility', key: 'facilityName', width: '24%' },
@@ -60,13 +59,11 @@ export default {
       cancelledApplicationId: undefined,
     }
   },
-  computed: {
-    ...mapState(useApplicationsStore, ['applications']),
-  },
+  computed: {},
   async created() {
     try {
       this.loading = true
-      await this.getApplications()
+      this.applications = await ApplicationService.getApplications()
     } catch (error) {
       this.setFailureAlert('Failed to get the list of applications')
     } finally {
@@ -74,7 +71,6 @@ export default {
     }
   },
   methods: {
-    ...mapActions(useApplicationsStore, ['getApplications']),
     isEmpty,
     getApplicationAction(application) {
       if (application?.status === 'Draft') {
@@ -92,8 +88,11 @@ export default {
       this.showCancelDialog = !this.showCancelDialog
       this.cancelledApplicationId = applicationId
     },
-    getRoutingPath(applicationId) {
-      return getApplicationUrl(PATHS.FACILITY_DETAILS, applicationId)
+    cancelApplication() {
+      const index = this.applications?.findIndex((item) => item.applicationId === this.cancelledApplicationId)
+      if (index > -1) {
+        this.applications.splice(index, 1)
+      }
     },
   },
 }
@@ -102,10 +101,5 @@ export default {
 <style scoped>
 :deep(div.v-data-table-footer) {
   padding-top: 20px;
-}
-:deep(.v-data-table-header__content:hover) {
-  font-weight: bold;
-  color: black;
-  cursor: pointer;
 }
 </style>
